+++ config/train_aws.yaml
@@ -7,3 +7,4 @@ epoch_check_point: 2
 batch_check_point: 20
 sample_text_len: 500
 early_stopping_criteria: 5
+seed: 123
