--- dvc.yaml
@@ -1,11 +1,0 @@
-stages:
-  preprocessing:
-    cmd: python bin/clean_annotations.py ./datasets/funsd/training_data/annotations
-      ./datasets/generated/training/annotations
-    deps:
-    - ./datasets/funsd/training_data/annotations
-    - bin/clean_annotations.py
-    params:
-    - preprocess.combined_fname
-    outs:
-    - ./datasets/generated/training/annotations
