--- model_artifacts/saved_model_003.dvc
@@ -1,5 +1,0 @@
-outs:
-- md5: 8aa12522e091268f53a3da532a7a201c.dir
-  size: 777035
-  nfiles: 4
-  path: saved_model_003
