--- model_artifacts/saved_model_004.dvc
@@ -1,5 +1,0 @@
-outs:
-- md5: 4079168305243efc965280a20daf258d.dir
-  size: 777035
-  nfiles: 4
-  path: saved_model_004
