--- model_artifacts/saved_model_006.dvc
@@ -1,5 +1,0 @@
-outs:
-- md5: 90c9b7a58c5c8f9751a997caa5e6079d.dir
-  size: 775762
-  nfiles: 4
-  path: saved_model_006
