--- model_artifacts/saved_model_006_wrapper.dvc
@@ -1,5 +1,0 @@
-outs:
-- md5: 48fb4acea3bf58fca2cf511c031eda0a.dir
-  size: 1644512
-  nfiles: 22
-  path: saved_model_006_wrapper
