--- model_artifacts/saved_model_baseline_wrapper.dvc
@@ -1,5 +1,0 @@
-outs:
-- md5: 0ecb96e47033e2786777ed3334a2fc4e.dir
-  size: 1644565
-  nfiles: 22
-  path: saved_model_baseline_wrapper
