+++ scripts/serve_model_local.sh
@@ -1,0 +1,1 @@
+mlflow models serve -p 5001 -m model_artifacts/saved_model_local_wrapper
